infrastructural change to convert image service to media service, acknolwedging that eventually you will be able to have multiple instances of a media object in the workspace, and it might contain different meta-data in each instance
